backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysql-connector-python
python-dotenv